+++ scripts/sql/profiles/h2-default/delta/v230-v240/apolloconfigdb-v230-v240.sql
@@ -33,6 +33,28 @@ CREATE ALIAS IF NOT EXISTS UNIX_TIMESTAMP FOR "com.ctrip.framework.apollo.common
 
 ALTER TABLE `AccessKey` ADD COLUMN `Mode` tinyint unsigned NOT NULL DEFAULT '0' COMMENT '密钥模式，0: filter，1: observer' AFTER `Secret`;
 
+ALTER TABLE `Commit` MODIFY COLUMN `ClusterName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Cluster Name';
+ALTER TABLE `Commit` MODIFY COLUMN `NamespaceName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Namespace Name';
+
+ALTER TABLE `Namespace` MODIFY COLUMN `ClusterName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Cluster Name';
+ALTER TABLE `Namespace` MODIFY COLUMN `NamespaceName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Namespace Name';
+
+ALTER TABLE `Release` MODIFY COLUMN `ClusterName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Cluster Name';
+ALTER TABLE `Release` MODIFY COLUMN `NamespaceName` VARCHAR(32) NOT NULL DEFAULT 'default' COMMENT 'Namespace Name';
+
+ALTER TABLE `Commit` DROP INDEX `Commit_ClusterName`;
+ALTER TABLE `Commit` DROP INDEX `Commit_NamespaceName`;
+ALTER TABLE `Commit` ADD INDEX `Commit_ClusterName`  (`ClusterName`);
+ALTER TABLE `Commit` ADD INDEX `Commit_NamespaceName`  (`NamespaceName`);
+
+ALTER TABLE `Namespace` DROP INDEX `Namespace_UK_AppId_ClusterName_NamespaceName_DeletedAt`;
+ALTER TABLE `Namespace` DROP INDEX `Namespace_IX_NamespaceName`;
+ALTER TABLE `Namespace` ADD UNIQUE INDEX `Namespace_UK_AppId_ClusterName_NamespaceName_DeletedAt`  (`AppId`,`ClusterName`,`NamespaceName`,`DeletedAt`);
+ALTER TABLE `Namespace` ADD INDEX `Namespace_IX_NamespaceName`  (`NamespaceName`);
+
+ALTER TABLE `Release` DROP INDEX `Release_AppId_ClusterName_GroupName`;
+ALTER TABLE `Release` ADD INDEX `Release_AppId_ClusterName_GroupName`  (`AppId`,`ClusterName`,`NamespaceName`);
+
 -- 
 -- ===============================================================================
 -- ==                                                                           ==